API reference

AtomsIO.AseParser — Type

Parse or write files using the Atomistic Simulation Environment (ASE)

AtomsIO.ChemfilesParser — Type

Parse or write file using Chemfiles

AtomsIO.ExtxyzParser — Type

Parse or write file using ExtXYZ

AtomsIO.load_system — Function

load_system([parser], file::AbstractString; kwargs...)
load_system([parser], file::AbstractString, index; kwargs...)

Read an AtomsBase-compatible system from file. If file contains more than one structure the last entry is returned. If index is specified this indexes into the list of structures and returns the respective system.

By default AtomsIO picks an appropriate parser for the specified file format automatically. A specific parser (such as ExtxyzParser or AseParser) can be enforced by using it as the first argument. Some parsers support additional keyword arguments. E.g. AseParser supports the format argument to overwrite the ASE-internal selection of an input / output format.

AtomsIO.load_trajectory — Method

load_trajectory([parser], file::AbstractString; kwargs...)

Read a trajectory from a file and return a vector of AtomsBase-compatible structures. Providing a parser overwrites the automatic AtomsIO selection.

AtomsIO.save_system — Method

save_system([parser], file::AbstractString, system::AbstractSystem; kwargs...)

Save an AtomsBase-compatible system to the file. By default AtomsIO picks an appropriate parser for the specified file format automatically. A specific parser (such as ExtxyzParser or AseParser) can be enforced by using it as the first argument. Some parsers support additional keyword arguments. E.g. AseParser supports the format argument to overwrite the ASE-internal selection of an input / output format.

AtomsIO.save_trajectory — Method

save_trajectory([parser], file::AbstractString, systems::AbstractVector; kwargs...)

Save a trajectory given as a list of AtomsBase-compatible systems to a file. Providing a parser overwrites the automatic AtomsIO selection.